Fix header issues
