+++ frontend/src/App.tsx
@@ -3,10 +3,8 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
 import { store } from "./store";
 
-import Body from "./components/home-page/body";
+import Home from "./components/home-page/home";
 import Gradebook from "./components/gradebook/gradebook";
-import HomeHeader from "./components/home-header/home-header";
-import PollHeader from "./components/poll-header/poll-header";
 import PollInProgress from "./components/poll-in-progress/poll-in-progress";
 import Present from "./components/present-poll/present";
 import Creator from "./components/creator-module/creator";
@@ -18,18 +16,8 @@ function App(): ReactElement {
   return (
     <Router>
       <Switch>
-        <Route path="/poll">
-          <PollHeader />
-        </Route>
-
-        <Route path="/">
-          <HomeHeader />
-        </Route>
-      </Switch>
-
-      <Switch>
         <Route exact path="/">
-          <Body />
+          <Home />
         </Route>
 
         <Route path="/poll/present">
