+++ frontend/src/components/home-page/content-tree.scss
@@ -1,7 +1,6 @@
 .content-tree {
   width: 40%;
   height: 100%;
-  // border: 1px solid red;
   border-right: 1px solid #cccccc;
 
   .tree-options {
@@ -10,13 +9,11 @@
     justify-content: space-evenly;
     width: 100%;
     height: 10%;
-    // border: 1px solid blue;
 
     .input-box {
       left: 0;
       border-radius: 8px;
       border-width: 1px;
-      // margin: 10px 10px;
       height: 50%;
       width: 70%;
       color: #a6a6a6;
@@ -42,11 +39,9 @@
   .question-list {
     width: 100%;
     height: 90%;
-    // border: 1px solid red;
 
     .question-list-header {
       height: 5%;
-      // border: 1px solid blue;
       display: flex;
       justify-content: space-between;
       align-items: center;
