+++ frontend/src/components/home-page/home.scss
@@ -3,5 +3,4 @@
   height: 85vh;
   width: 100%;
   display: flex;
-  // border: 1px solid blue;
 }
