+++ frontend/src/components/home-page/home.tsx
@@ -1,14 +1,20 @@
 import React, { ReactElement } from "react";
 import ContentTree from "./content-tree";
 import QuestionPreview from "./question-preview";
-import "./body.scss";
+import "./home.scss";
+import HomeHeader from "../home-header/home-header";
 
 const Body = (): ReactElement => {
   return (
-    <div className="body">
-      <ContentTree />
-      <QuestionPreview />
-    </div>
+    <>
+      <HomeHeader />
+
+      <div className="body">
+        <ContentTree />
+
+        <QuestionPreview />
+      </div>
+    </>
   );
 };
