+++ frontend/src/components/poll-header/poll-header.scss
@@ -1,6 +1,4 @@
 .poll-header {
-  height: 7%;
-  // border: 1px solid red;
   background: black;
 
   display: flex;
@@ -32,6 +30,5 @@
     border-radius: 5px;
     border: 0px;
     background: #ffc904;
-    // padding: 1vh 2vw;
   }
 }
