--- frontend/src/components/present-poll/present-footer.scss
@@ -1,32 +1,0 @@
-.present-footer {
-  height: 7%;
-  background: black;
-
-  display: flex;
-  align-items: stretch;
-  overflow-x: scroll;
-
-  .footer-question {
-    background: white;
-    display: flex;
-    align-items: center;
-    justify-content: space-evenly;
-    margin: 0.25vh 0.15vw;
-    min-width: 20vw;
-
-    img {
-      height: 3.5vh;
-      width: auto;
-    }
-
-    p {
-      width: 55%;
-    }
-
-    .delete {
-      border: none;
-      background: white;
-      color: #aaaaaa;
-    }
-  }
-}
