+++ frontend/src/components/present-poll/present-footer/present-footer.scss
@@ -1,0 +1,7 @@
+.present-footer {
+  background: black;
+
+  display: flex;
+  flex-shrink: 0;
+  overflow-x: scroll;
+}
