+++ frontend/src/components/present-poll/present-preview/present-preview.scss
@@ -1,5 +1,4 @@
 .present-preview {
-  // border: 1px solid blue;
   height: 66vh;
   padding: 10vh 10vw;
 
@@ -9,7 +8,6 @@
   align-items: flex-start;
 
   .present-preview-logo {
-    // border: 1px solid red;
     height: 10vh;
     width: auto;
   }
