+++ frontend/src/components/present-poll/present-preview/present-preview.tsx
